moved faqsection to wihe dir
